registry: move search-related code to separate files
Also touching-up some comments.
