+++ registry/search_endpoint_v1_test.go
