+++ registry/search_session.go
